thdat: Rewrite.
Generate a shared library containing the thdat functions.
Add an IO library.
Add an error library.
Add a command for listing the conents of archives.
Provide a pkg-config file for the shared library.
Rework libraries to use the IO library for IO.
Dependencies for the library have been kept small.